requirements.txt
# ============================================================
# 生田緑地プロジェクト requirements.txt
# 確定日：2026-02-28  Python 3.11 以上
# ============================================================

# ── データ取得 ──────────────────────────────────────────────
pystac-client>=0.8,<1.0
planetary-computer>=1.0,<2.0

# ── STACアイテム → xarray 変換（メインエンジン）──────────
# stackstac は Inactive のため不採用。odc-stac に一本化。
odc-stac>=0.5.2,<1.0

# ── データ処理・配列操作 ───────────────────────────────────
xarray>=2024.6,<2026.0
rioxarray>=0.18,<1.0
numpy>=1.26,<3.0
dask[complete]>=2024.5,<2026.0

# ── COG生成 ───────────────────────────────────────────────
rio-cogeo>=5.0,<8.0

# ── 地理座標系変換 ─────────────────────────────────────────
pyproj>=3.6,<4.0

# ── CSV/JSON出力 ───────────────────────────────────────────
pandas>=2.0,<3.0

# ── リトライ処理 ───────────────────────────────────────────
tenacity>=8.2,<10.0
